Создал модульность кода
