+++ sanitizeHtml.js
@@ -1,0 +1,5 @@
+function sanitizeHtml(htmlString = "") {
+    return htmlString.replaceAll("<", "&lt;").replaceAll(">", "&gt;");
+  }
+
+  export {sanitizeHtml};
